Fix #14: Allows relative imports with subdirectories.
If a and b are siblings, and you are importing `foo` from b into a,
before this change, you'd import like `from . import foo`.

However, if a has a child namespace `bar`, then things in a will live in
an `a/__init__.py` file, so the import needs to be `from .. import foo`.

I fixed this by making the `bridge_namespaces` method no longer
static (as the object it gets called on has the context of other
namespaces) and then checking if the import-to namespace is a parent of
another namespace.

If you would rather we pass the namespaces as an argument or a flag we
culd do that, but since the only actual call has the object, it felt
reasonable to do it this way (it did make the test a bit more awkward).

I have also added a couple of tests.

See: #14