Update tests to avoid issue with union of record validation.
In the newest version of avro there appears to be a bug when dealing
with a schema which is a union of records that have the same fields with
different types.

See: #16

I have update the example schema so it doesn't have that problem, to
avoid running into this issue.

This doesn't _solve_ the issue per se, but it avoids it.

Also, I changed the strategy for generating decimals to not generate as
many decimals we don't care about in the test, since it was frustrating
to debug with it generating so many bad test cases we would abandon.